requirements.txt
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-Mail==0.10.0
werkzeug==3.1.3
openai==0.28